Move DownloadServiceConfig in own package and rename all References
