+++ services/download/src/test/java/app/coronawarn/server/services/download/FederationBatchProcessorTest.java
@@ -41,7 +41,8 @@ import app.coronawarn.server.common.persistence.domain.FederationBatchInfo;
 import app.coronawarn.server.common.persistence.domain.FederationBatchStatus;
 import app.coronawarn.server.common.persistence.service.DiagnosisKeyService;
 import app.coronawarn.server.common.persistence.service.FederationBatchInfoService;
-import app.coronawarn.server.services.download.DownloadServiceConfig.TekFieldDerivations;
+import app.coronawarn.server.services.download.config.DownloadServiceConfig;
+import app.coronawarn.server.services.download.config.DownloadServiceConfig.TekFieldDerivations;
 import app.coronawarn.server.services.download.validation.ValidFederationKeyFilter;
 import feign.FeignException;
 import java.time.LocalDate;
