+++ services/download/src/test/java/app/coronawarn/server/services/download/FederationBatchTestHelper.java
@@ -25,6 +25,7 @@ import static org.mockito.Mockito.when;
 
 import app.coronawarn.server.common.protocols.external.exposurenotification.DiagnosisKey;
 import app.coronawarn.server.common.protocols.external.exposurenotification.DiagnosisKeyBatch;
+import app.coronawarn.server.services.download.config.DownloadServiceConfig;
 import app.coronawarn.server.services.download.normalization.FederationKeyNormalizer;
 import com.google.protobuf.ByteString;
 import java.util.Optional;
